split backend_gtk3agg.py into two parts, the gcf dependent and gcf independent parts.
Same poor naming scheme as everything else in this branch.
